fix login seasion

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('NIK')->unique();
+            $table->bigInteger('NIK')->unique();
             $table->string('email')->unique();
             $table->string('phone')->nullable()->unique();
             $table->enum('role', ['admin', 'user'])->default('user');
